backend/requirements.txt
fastapi==0.114.2
uvicorn[standard]==0.30.6
SQLAlchemy==2.0.36
asyncpg==0.29.0
pydantic==2.9.2
pydantic-settings==2.4.0
httpx==0.27.2
beautifulsoup4==4.12.3
python-dotenv==1.0.1
openai==1.11.0
